% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plotting Accidents on Map for Given State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric/integer State Code
\tabular{cc}{
  \strong{State Code} \tab \strong{State Name}    \cr
  01 \tab  Alabama              \cr
  02 \tab  Alaska               \cr
  04 \tab  Arizona              \cr
  05 \tab  Arkansas             \cr
  06 \tab  California           \cr
  08 \tab  Colorado             \cr
  09 \tab  Connecticut          \cr
  10 \tab  Delaware             \cr
  11 \tab  District of Columbia \cr
  12 \tab  Florida              \cr
  13 \tab  Georgia              \cr
  15 \tab  Hawaii               \cr
  16 \tab  Idaho                \cr
  17 \tab  Illinois             \cr
  18 \tab  Indiana              \cr
  19 \tab  Iowa                 \cr
  20 \tab  Kansas               \cr
  21 \tab  Kentucky             \cr
  22 \tab  Louisiana            \cr
  23 \tab  Maine                \cr
  24 \tab  Maryland             \cr
  25 \tab  Massachusetts        \cr
  26 \tab  Michigan             \cr
  27 \tab  Minnesota            \cr
  28 \tab  Mississippi          \cr
  29 \tab  Missouri             \cr
  30 \tab  Montana              \cr
  31 \tab  Nebraska             \cr
  32 \tab  Nevada               \cr
  33 \tab  New Hampshire        \cr
  34 \tab  New Jersey           \cr
  35 \tab  New Mexico           \cr
  36 \tab  New York             \cr
  37 \tab  North Carolina       \cr
  38 \tab  North Dakota         \cr
  39 \tab  Ohio                 \cr
  40 \tab  Oklahoma             \cr
  41 \tab  Oregon               \cr
  42 \tab  Pennsylvania         \cr
  43 \tab  Puerto Rico          \cr
  44 \tab  Rhode Island         \cr
  45 \tab  South Carolina       \cr
  46 \tab  South Dakota         \cr
  47 \tab  Tennessee            \cr
  48 \tab  Texas                \cr
  49 \tab  Utah                 \cr
  50 \tab  Vermont              \cr
  51 \tab  Virginia             \cr
  52 \tab  Virgin Islands       \cr
  53 \tab  Washington           \cr
  54 \tab  West Virginia        \cr
  55 \tab  Wisconsin            \cr
  56 \tab  Wyoming
}}

\item{year}{A numeric or character value for a particular Year}
}
\value{
The map of accidents for the entered State and Year
}
\description{
This is a function that takes a particular State
as an Input along with a particular Year and plots
the accidents on the map of that State. It returns
and error if the State provided is invalid and also
returns a blank plot if there are no accidents to
plot
}
\examples{
\dontrun{
fars_map_state(49,2015)
}
}
