% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract Data for Given Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of numeric or character values of years}
}
\value{
The data frame consisting of all values of months and years, else NULL
}
\description{
This is a function that reads all data for the
given vector of years and returns a warning if
any of the entered years is not present in the
data. The function also makes prints separate
filenames for every year by calling the function
make_filename
}
\examples{
fars_read_years(2013)
fars_read_years(c(2013,2014))
}
