% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarising Data for Years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of numeric or character values of years}
}
\value{
The data frame consisting of count of accidents by month and year
}
\description{
This is a function that reads all data for the
given vector of years and groups the data by
month and year - one record for every month and
year combination
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))
}
