% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{File Name Printing Function for a Particular Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric or character value for the year}
}
\value{
The file name for the respective year in zipped-csv format
}
\description{
This is a function that prints the respective filename
for the \code{year} value provided
}
\examples{
make_filename(2013)
make_filename(2014)
}
