% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plotting Accidents on Map for Given State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric/integer State Code}

\item{year}{A numeric or character value for a particular Year}
}
\value{
The map of accidents for the entered State and Year
}
\description{
This is a function that takes a particular State
as an Input along with a particular Year and plots
the accidents on the map of that State. It returns
and error if the State provided is invalid and also
returns a blank plot if there are no accidents to
plot
}
\examples{
fars_map_state(49,2015)
}
