% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{File Reading Function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A name for the file to be read}
}
\value{
The data frame created after reading the file \code{filename}
}
\description{
This is a function that first checks if a file exists
otherwise uses the readr package's read_csv function
to read the file suppressing the messages during that
process. Finally it converts the data read above into
a data frame
}
\details{
For more information, see:
\itemize{
  \item{\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}}
  \item{\url{https://en.wikipedia.org/wiki/Fatality_Analysis_Reporting_System}}
}
}
\note{
To generate file name use: \code{\link{make_filename}}
}
\examples{
library(dplyr)
library(readr)
yr <- 2015
data <- yr \%>\%
  make_filename \%>\%
  fars_read
head(data)
}
\seealso{
\link{make_filename}
}
